return float result

+++ app.js
@@ -2,7 +2,7 @@
 var FFI = require('ffi');
 
 var lib = FFI.Library('./target/debug/liblibshingles.so', {
-    'count_same': [ 'int', [ 'string', 'string', 'int' ] ]
+    'count_same': [ 'float', [ 'string', 'string', 'int' ] ]
 });
 
 module.exports = lib.count_same;
